requirements.txt
PyQt6
numpy
opencv-python
requests
openpyxl
pillow
pytesseract 